Fix sync feature.
